Refactored Code Structure. Fixed All Paths.
